feat: replace `RawCatalogSearchResultFormatterPlugin`-overwrite
- implement plugin contract `ProductExpanderPluginInterface` from bundle `fond-of-impala/enhanced-catalog` for preparing grouped product image data
- remove `RawCatalogSearchResultFormatterPlugin`-overwrite
- cleanup dependencies
  - remove "spryker/catalog": "^3.0.0 || ^4.0.0 || ^5.0.0",
  - add "spryker/enhanced-catalog-extension": "^1.0.0",
- ...
