+++ bundles/product-image-grouping-product-page-search/composer.json
@@ -10,7 +10,7 @@
   ],
   "require": {
     "php": ">=8.0",
-    "spryker/catalog": "^3.0.0 || ^4.0.0 || ^5.0.0",
+    "spryker/enhanced-catalog-extension": "^1.0.0",
     "spryker/product-page-search": "^3.24.0",
     "spryker/product-page-search-extension": "^1.4.0"
   },
