+++ dandelion.json
@@ -321,7 +321,7 @@
     },
     "product-image-grouping-product-page-search": {
       "path": "bundles/product-image-grouping-product-page-search",
-      "version": "1.2.0"
+      "version": "1.3.0"
     },
     "product-list-conditional-availability-page-search": {
       "path": "bundles/product-list-conditional-availability-page-search",
